pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeforge"
version = "0.1.0"
description = "Coding-focused LLM built from scratch with PyTorch and TCFP (Tensor Core Floating Point)"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [{name = "bwiemz"}]
keywords = [
    "llm", "code-generation", "transformer", "fp8",
    "tensor-cores", "quantization", "deep-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.4.0",
    "tokenizers>=0.19.0",
    "datasets>=2.14.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
triton = ["triton>=3.0.0"]
tracking = ["wandb>=0.15.0"]
quantize = ["bitsandbytes>=0.41.0"]
dev = ["pytest>=7.0", "ruff>=0.4.0", "basedpyright>=1.20.0"]
all = ["codeforge[triton,tracking,quantize,dev]"]

[project.scripts]
codeforge-train = "scripts.train:main"
codeforge-generate = "scripts.generate:main"

[tool.setuptools.packages.find]
include = ["codeforge*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
reportMissingImports = "none"
reportAny = "none"

[tool.pytest.ini_options]
testpaths = ["tests"]
